Refactor storage and sync monoliths to reduce complexity
