+++ server/constants/platforms.ts
@@ -1,0 +1,12 @@
+export const popularStreamingPlatforms: Record<string, number> = {
+  Netflix: 203,
+  'Amazon Prime Video': 26,
+  Hulu: 157,
+  'HBO Max': 384,
+  'Disney+': 372,
+  'Apple TV+': 371,
+  'Paramount+': 389,
+  Peacock: 386,
+  Shudder: 99,
+  Tubi: 283,
+};
